gh-680: extract janusgraph connector state into subpackage
This seemingly cosmetic change is required for us to build a filter
subpackage. Otherwise we'll run into circular dependencies because of
the MappedClassName and MappedProperty types and their related methods.
