+++ database/connectors/janusgraph/helper.go
@@ -17,6 +17,7 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/creativesoftwarefdn/weaviate/database/connectors/janusgraph/state"
 	connutils "github.com/creativesoftwarefdn/weaviate/database/connectors/utils"
 	"github.com/creativesoftwarefdn/weaviate/database/schema"
 	"github.com/creativesoftwarefdn/weaviate/database/schema/kind"
@@ -71,8 +72,8 @@ func (j *Janusgraph) getClass(k kind.Kind, searchUUID strfmt.UUID, atClass *stri
 	}
 
 	kind := kind.KindByName(vertex.AssertPropertyValue(PROP_KIND).AssertString())
-	mappedClassName := MappedClassName(vertex.AssertPropertyValue(PROP_CLASS_ID).AssertString())
-	className := j.state.getClassNameFromMapped(mappedClassName)
+	mappedClassName := state.MappedClassName(vertex.AssertPropertyValue(PROP_CLASS_ID).AssertString())
+	className := j.state.GetClassNameFromMapped(mappedClassName)
 	class := j.schema.GetClass(kind, className)
 	if class == nil {
 		panic(fmt.Sprintf("Could not get %s class '%s' from schema", kind.Name(), className))
@@ -100,8 +101,8 @@ func (j *Janusgraph) getClass(k kind.Kind, searchUUID strfmt.UUID, atClass *stri
 	// Just copy in the value directly. We're not doing any sanity check/casting to proper types for now.
 	for key, val := range vertex.Properties {
 		if strings.HasPrefix(key, "prop_") {
-			mappedPropertyName := MappedPropertyName(key)
-			propertyName := j.state.getPropertyNameFromMapped(className, mappedPropertyName)
+			mappedPropertyName := state.MappedPropertyName(key)
+			propertyName := j.state.GetPropertyNameFromMapped(className, mappedPropertyName)
 			err, property := j.schema.GetProperty(kind, className, propertyName)
 			if err != nil {
 				panic(fmt.Sprintf("Could not get property '%s' in class %s ; %v", propertyName, className, err))
@@ -125,10 +126,10 @@ func (j *Janusgraph) getClass(k kind.Kind, searchUUID strfmt.UUID, atClass *stri
 	for _, edge := range refEdges {
 		locationUrl := edge.AssertPropertyValue(PROP_REF_EDGE_LOCATION).AssertString()
 		type_ := edge.AssertPropertyValue(PROP_REF_EDGE_TYPE).AssertString()
-		mappedPropertyName := MappedPropertyName(edge.AssertPropertyValue(PROP_REF_ID).AssertString())
+		mappedPropertyName := state.MappedPropertyName(edge.AssertPropertyValue(PROP_REF_ID).AssertString())
 		uuid := edge.AssertPropertyValue(PROP_REF_EDGE_CREF).AssertString()
 
-		propertyName := j.state.getPropertyNameFromMapped(className, mappedPropertyName)
+		propertyName := j.state.GetPropertyNameFromMapped(className, mappedPropertyName)
 		err, property := j.schema.GetProperty(kind, className, propertyName)
 		if err != nil {
 			panic(fmt.Sprintf("Could not get property '%s' in class %s ; %v", propertyName, className, err))
@@ -181,11 +182,11 @@ func (j *Janusgraph) listClass(k kind.Kind, className *schema.ClassName, first i
 		// filter spike
 		filterProp := filters.Root.On.Property
 		filterValue := filters.Root.Value.Value
-		q = q.Raw(fmt.Sprintf(".has(\"%s\", lt(%d))", j.state.getMappedPropertyName(*className, filterProp), filterValue))
+		q = q.Raw(fmt.Sprintf(".has(\"%s\", lt(%d))", j.state.GetMappedPropertyName(*className, filterProp), filterValue))
 	}
 
 	if className != nil {
-		vertexLabel := j.state.getMappedClassName(*className)
+		vertexLabel := j.state.GetMappedClassName(*className)
 		q = q.HasString(PROP_CLASS_ID, string(vertexLabel))
 	}
